api/pyproject.toml
[project]
name = "trader-api"
version = "1.0.0"
description = "Trading recommendation REST API with PostgreSQL backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "yfinance>=0.2.36",
    "anthropic>=0.45",
    "pandas>=2.2",
    "numpy>=1.26",
    "ta-lib>=0.5",
    "pyyaml>=6.0",
    "fear-greed>=0.1",
    "python-multipart>=0.0.9",
    "PyJWT[crypto]>=2.8",
    "httpx[http2]>=0.27",
    "py-webauthn>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
    "httpx>=0.27",
]

[project.scripts]
trader-api = "trader_api.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
